04_resume_fit_analyzer/requirements.txt
streamlit
langchain
openai
python-dotenv
PyPDF2
